+++ sim/game/cards/dm10/chimera.go
@@ -47,3 +47,17 @@ func Gigandura(c *match.Card) {
 	}))
 
 }
+
+// Gigamente ...
+func Gigamente(c *match.Card) {
+
+	c.Name = "Gigamente"
+	c.Power = 3000
+	c.Civs = []string{civ.Darkness}
+	c.Family = []string{family.Chimera}
+	c.ManaCost = 4
+	c.ManaRequirement = []string{civ.Darkness}
+
+	c.Use(fx.Creature, fx.SilentSkill(fx.ReturnCreaturesFromGraveToHand(1, false)))
+
+}