+++ sim/game/cards/dm10/soltrooper.go
@@ -57,3 +57,17 @@ func PoltalesterTheSpydroid(c *match.Card) {
 	c.Use(fx.Creature, fx.Blocker(), fx.ShieldTrigger, fx.CantAttackPlayers)
 
 }
+
+// BulglufTheSpydroid ...
+func BulglufTheSpydroid(c *match.Card) {
+
+	c.Name = "Bulgluf, the Spydroid"
+	c.Power = 4000
+	c.Civs = []string{civ.Light}
+	c.Family = []string{family.Soltrooper}
+	c.ManaCost = 6
+	c.ManaRequirement = []string{civ.Light}
+
+	c.Use(fx.Creature, fx.SilentSkill(fx.TopCardToShield))
+
+}